Stability and hearbeat enhancements
*) Changed the connected signal to be readyToSend

*) Changed the disconnected signal to be notReadyToSend

*) Implemeted the heartbeat functionality by adding a
heartbeat_interval settings option.  If set to 0 hearbeat
is disabled, grater than 0 sets the heartbeat to that many
seconds. Exposed two new signals heartbeat and notHeartbeat so
the Qml viewer is hearbeat aware.  Added two slots so the Qml
viewer can enable and disable the heartbeat on while running.

*) Updated settings.conf.example to reflect new heartbeat setting

*) Fixed messageHandler parsing bug where a certain message caused
the Qml viewer to segfault.  The messageHandler item and value lists
were not checked for the correct size.

Signed-off-by: Jeff Horn <jhorn@reachtech.com>